moves deliveryservice list view into its own namespace
