+++ traffic_ops/experimental/ui/app/src/common/modules/navigation/navigation.tpl.html
@@ -15,7 +15,7 @@
                 </li>
                 <li class="side-menu-category"><a href="javascript:void(0);"><i class="fa fa-cog"></i> Configure <span class="fa fa-chevron-down"></span></a>
                     <ul class="nav child_menu" style="display: none">
-                        <li class="side-menu-category-item" ng-class="{'current-page': isState('trafficOps.private.configure.deliveryServices')}"><a ng-click="navigateToPath('/configure/delivery-services')">Delivery Services</a></li>
+                        <li class="side-menu-category-item" ng-class="{'current-page': isState('trafficOps.private.configure.deliveryServices.list')}"><a ng-click="navigateToPath('/configure/delivery-services')">Delivery Services</a></li>
                     </ul>
                 </li>
                 <li class="side-menu-category"><a href="javascript:void(0);"><i class="fa fa-wrench"></i> Administer <span class="fa fa-chevron-down"></span></a>
