+++ traffic_ops/experimental/ui/app/src/modules/private/configure/deliveryServices/index.js
@@ -4,18 +4,13 @@ module.exports = angular.module('trafficOps.private.configure.deliveryServices',
         $stateProvider
             .state('trafficOps.private.configure.deliveryServices', {
                 url: '/delivery-services',
+                abstract: true,
                 views: {
                     configureContent: {
                         templateUrl: 'modules/private/configure/deliveryServices/deliveryServices.tpl.html',
                         controller: 'DeliveryServicesController'
                     }
-                },
-                resolve: {
-                    deliveryServices: function(user, deliveryServiceService) {
-                        return deliveryServiceService.getDeliveryServices(false);
-                    }
                 }
-
             })
         ;
         $urlRouterProvider.otherwise('/');
